CTCP-517: Add connection to transit-movements-router
There were also a few updates to existing v2 code to enable easier testing of various components. Of note, "MessageSizeAction" should now be injected via a provider due to turning it into a trait and the generic signature and need to fake out the action. The AuthenticationOnlyNewEnrolment action was also split into a trait to make faking the authentication easier.