CTCP-773: Fix routing to V1 GET/DELETE endpoints when a content type is defined
If we directly invoke the action via a GET endpoint with a content type, Play correctly parses the body as empty. However, it seems when chaining the actions, Play will use the content type and try to parse an empty body. To sidestep this, we drain the request body (which should be empty anyway) and manually remove the content type so that the GET endpoints don't try to parse things as the content type suggests.