CTCP-2295: Move from hybrid memory/file streaming to file based streaming
This has been necessitated as we've been finding that our code has been too fast for Akka occasionally, as it might start to execute a step when the stream to file hasn't finished, even through we put in safeguards to prevent that. It was a nice experiment, but we need to value reliability over speed.